Fixes response handling
The callbacks for directing the next action are mangled in such a way that:

Requesting response inspection will never result in a response inspection callback.
Request headers are sent to the callbacks for the response messages.
We completely ignore the proxy next action directive on responses.

All issues fixed here.

Fixes #11