Reorder websocket negotiation
Reorders websocket negotiation to get the negotiated subprotocol from the upstream websocket server before accepting the incoming client connection, in order to properly pass this on to the connected client and thus synchronize the subprotocol between upstream and downstream.

Possible fix but probably not quite there yet, untested.
Related to #27