Remove coverity, clean up HTTP header checks
When handling a websocket connection, we were looking to the list of forbidden HTTP headers being returned when deciding which headers to forward on a websocket connection. To avoid confusion and, to avoid issues of overlap in the protocol handling, we move over initial things that we know for sure we'll want to block to the forbidden websocket headers list, and stop looking at the forbidden http headers list.

We may need to add more headers now that we've changed this over.

Remove coverity scan because it doesn't work presently. We just do the builds manually.