2.x API, new features, fixes
The engine now defines a true firewall API, where clients can dictate which applications should be allowed to access the internet depending on port number, and also whether or not to filter the applications, so on and so forth. However, this needs to be implemented by concrete (platform-specific) implementations.

Adds response compression which is always on now. This way the engine can be used in a remote proxy as well as a local proxy.

Fixes two issues identified by coverity.

Enables users to specify their own common name for the spoofed certificate authority.