Fixes
Upgrades bouncy castle and kestrel.
Fixes an issue where the status code property of the message info object is not initialized correctly.
Documents the fact that modifications to the message info class brings the burden of ensuring the validity of the object entirely on to the shoulders of the user.

Fixes #62
Fixes #59