Fixes
Changes pub API by adding the source message info object to the replay callback handler. This is for the sake of connection tracking outside the engine.
Changed URL retrieval to be non-encoded. I think this could cause issues as the underlying mechanics in .NET might (maybe?) double-encode our URL's without this.
Fixed an issue with websockets.

Fixes #71